+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.4 (unreleased)
 
+- Fixed ORM methods for datafiles with relative path patterns.
 - Added plugin for `mypy` support.
 
 # 0.3 (2019-06-09)