+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.4 (unreleased)
+# 0.4 (2019-06-29)
 
 - Fixed ORM methods for datafiles with relative path patterns.
 - Added plugin for `mypy` support.